Systems can now be run in any environment not only production
