+++ foreman/.env
@@ -1,3 +1,4 @@
+RAILS_ENV=production
 DB_HOST=${DB_HOST}
 DB_USERNAME=postgres
 REDIS_HOST=${REDIS_HOST}
@@ -14,7 +15,6 @@ AWS_S3_FORCE_PATH_STYLE='true'
 RAILS_SERVE_STATIC_FILES='true'
 ATOS_API_URL="http://atos.et.127.0.0.1.nip.io:3100"
 ACAS_API_URL="http://api.et.127.0.0.1.nip.io:3100/et_acas_api"
-ET_API_URL='http://api.et.127.0.0.1.nip.io:3100/api/v2'
 ATOS_API_USERNAME='atos'
 ATOS_API_PASSWORD='password'
 RAILS_MAX_THREADS=5
