+++ foreman/Procfile
@@ -2,7 +2,7 @@ traefik: bash --login -c "traefik --file.filename=${FOREMAN_PATH}/traefik.toml"
 et1_web:  bash --login -c "cd ${FS_ROOT_PATH} && et_full_system local wait_for_support && et_full_system local update_service_url et1 http://localhost:${PORT} && cd systems/et1 && dotenv -f ${FOREMAN_PATH}/.env dotenv -f ${FOREMAN_PATH}/et1.env ./run.sh"
 et1_sidekiq: bash --login -c "cd ${FS_ROOT_PATH} && et_full_system local wait_for_support && cd systems/et1  && dotenv -f ${FOREMAN_PATH}/.env dotenv -f ${FOREMAN_PATH}/et1.env ./run_sidekiq.sh"
 et3_web:  bash --login -c "cd ${FS_ROOT_PATH} && et_full_system local wait_for_support && et_full_system local update_service_url et3 http://localhost:${PORT} && cd systems/et3 && dotenv -f ${FOREMAN_PATH}/.env dotenv -f ${FOREMAN_PATH}/et3.env ./run.sh"
-mail_web: bash --login -c "source ./mailhog.env && mailhog"
+mail_web: bash --login -c "source ./mailhog.env && mailhog &> /dev/null"
 api_web: bash --login -c "cd ${FS_ROOT_PATH} && et_full_system local wait_for_support && et_full_system local update_service_url api http://localhost:${PORT} && cd systems/api && dotenv -f ${FOREMAN_PATH}/.env dotenv -f ${FOREMAN_PATH}/et_api.env ./run.sh"
 api_sidekiq: bash --login -c "cd ${FS_ROOT_PATH} && et_full_system local wait_for_support && cd systems/api && dotenv -f ${FOREMAN_PATH}/.env dotenv -f ${FOREMAN_PATH}/et_api.env ./run_sidekiq.sh"
 admin_web: bash --login -c "cd ${FS_ROOT_PATH} && et_full_system local wait_for_support && et_full_system local update_service_url admin http://localhost:${PORT} && cd systems/admin && dotenv -f ${FOREMAN_PATH}/.env dotenv -f ${FOREMAN_PATH}/et_admin.env ./run.sh"
