+++ foreman/et1.env
@@ -1,4 +1,3 @@
-RAILS_ENV=local
 DB_NAME=et1db
 HTTP_OPEN_TIMEOUT_S=15
 HTTP_READ_TIMEOUT_S=15
@@ -11,4 +10,7 @@ S3_UPLOAD_BUCKET=et1bucket
 AWS_REGION=us-east-1
 AWS_ENDPOINT=http://s3.et.127.0.0.1.nip.io:3100
 AWS_S3_FORCE_PATH_STYLE=true
-
+ET_API_URL='http://api.et.127.0.0.1.nip.io:3100/api/v2'
+SECURE_SESSION_COOKIE=false
+SENDING_HOST=et1.et.127.0.0.1.nip.io
+FLATTEN_PDF=false
