+++ foreman/et3.env
@@ -1,4 +1,3 @@
-RAILS_ENV=production
 REDIS_URL=redis://redis:6379/5
 RAVEN_DSN='https://7f2374c2e0cd4650bf5129c9fe45215e:e0ae53d01bd7454da963790ee204ffa0@sentry.io/1376078'
 ET_API_URL='http://api.et.127.0.0.1.nip.io:3100/api'
