+++ foreman/et_admin.env
@@ -1,4 +1,3 @@
-RAILS_ENV=production
 DEVISE_SECRET_KEY=JK3N45KH56743JN25J1B5H43B24H35BJ6575KN21NRNKFASDF89DU23I4NK5TGKFD
 REDIS_DATABASE=2
 RAILS_SERVE_STATIC_FILES='true'
