+++ foreman/et_api.env
@@ -1,4 +1,3 @@
-RAILS_ENV=production
 WEB_CONCURRENCY=1
 SECRET_KEY_BASE=sdlkjewfkjhfviuhduihenrjk435r89esfd7cv983qh2n4r3q27yh4u5rtfg
 REDIS_DATABASE=2
