+++ foreman/traefik.toml
@@ -36,24 +36,3 @@
   debug = true
 
 [rest]
-    # Backends
-    [backends]
-
-      [backends.et1backend]
-
-        [backends.et1backend.servers]
-          [backends.et1backend.servers.rails]
-            url = "http://localhost:5000"
-            weight = 1
-
-    # Frontends
-    [frontends]
-
-      [frontends.et1frontend]
-        entryPoints = ["http"]
-        backend = "et1backend"
-        passHostHeader = true
-        priority = 42
-        [frontends.et1frontend.routes.test_1]
-            rule = "HostRegexp:localhost,{subdomain:[a-z]+}.localhost"
-            backend = "et1backend"
